requirements.txt
aiogram
aiohttp
aiosignal
async-timeout
attrs
Babel
certifi
cffi
charset-normalizer
colorama
docopt
frozenlist
idna
multidict
num2words
numpy==1.24.1
pycparser
python-dotenv
pytz
requests
srt
torch
tqdm
typing_extensions
urllib3
vosk
websockets
yarl
g4f[all]
PyPDF2
python-docx